split render.mts into multiple files
