+++ src/alias.mts
@@ -1,4 +1,4 @@
-import { createRenderer } from "./render.mjs";
+import { createRenderer } from "./renderer.mjs";
 import { LagopusObjectData, LagopusObjectOptions, LagopusGroup } from "./primes.mjs";
 
 export let group = (options: any, ...children: any[]): LagopusGroup => {
