+++ src/index.mts
@@ -1,6 +1,8 @@
 export { group, object, u32buffer, newBufferFormatLength, newBufferFormatArray } from "./alias.mjs";
 
-export { createRenderer, paintLagopusTree, renderLagopusTree, resetCanvasSize, createTextureFromSource } from "./render.mjs";
+export { paintLagopusTree, renderLagopusTree, resetCanvasSize } from "./paint.mjs";
+
+export { createRenderer } from "./renderer.mjs";
 
 export { initializeContext, enableBloom, initializeCanvasTextures } from "./initialize.js";
 
@@ -9,3 +11,5 @@ export { compButton, compDragPoint, compSlider } from "./comp/button.mjs";
 export { onControlEvent, registerShaderResult, loadGamepadControl } from "./control.mjs";
 
 export { setupMouseEvents } from "./events.mjs";
+
+export { createTextureFromSource } from "./util.mjs";
