+++ src/main.mts
@@ -1,6 +1,8 @@
 import queryString from "query-string";
-import { createTextureFromSource, paintLagopusTree, renderLagopusTree, resetCanvasSize } from "./render.mjs";
+import { paintLagopusTree } from "./paint.mjs";
+import { renderLagopusTree, resetCanvasSize } from "./renderer.mjs";
 import { enableBloom, initializeCanvasTextures, initializeContext } from "./initialize.js";
+import { createTextureFromSource } from "./util.mjs";
 
 import { compContainer } from "./app/container.mjs";
 import { renderControl, startControlLoop } from "@triadica/touch-control";
